src/requirements.txt
Flask==3.0.3
mariadb
pymssql
py-healthcheck
pysftp
psycopg2
prometheus-client
python-dotenv
requests==2.31.0
requests-pkcs12==1.24
SQLAlchemy